+++ ClientDashboard_API/DTOs/RevenuePatternsDto.cs
@@ -11,8 +11,8 @@ namespace ClientDashboard_API.DTOs
 
         public decimal TotalRevenue { get; set; }
 
-        public double RevenuePerWorkingDay { get; set; }
+        public decimal RevenuePerWorkingDay { get; set; }
 
-        public double RevenuePerWorkingWeek { get; set; }
+        public decimal RevenuePerWorkingWeek { get; set; }
     }
 }
